refactor: MainView store로 전역 store 폴더와 분리
