+++ src/components/mainView/store/reducer.ts
